Optimize TaskStateCounts aggregate pipeline
In its current state, MongoDB does not use indices in $group operations within an aggregate pipeline.  If a lot of tasks exist, this means that the simple aggregate sum operation will take a long time as it has to access all documents in the db and cannot just use the index of the state field. In real world usage, this was the longest frequently running query and caused significant load on our db. However, $match and $sort both are capable of using said index, and inserting them at the beginning of the pipeline yields a roughly ten-fold decrease in query runtime (in our hands) as $group can then operate on the $match or $sort results rather than polling all documents.